+++ spec/recipes/OneDriver/allocate_one_vm.rb
@@ -12,20 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'chef/provisioning/opennebula_driver'
-require 'fileutils'
-
-FileUtils.mkdir_p(CHEF_REPO_PATH.chomp('/'))
-with_chef_local_server :chef_repo_path => CHEF_REPO_PATH.chomp('/')
-with_driver DRIVER_URL
-with_flow_url ONE_FLOW_URL.chomp('/') unless ONE_FLOW_URL.nil?
-
-with_machine_options MACHINE_OPTIONS.merge(
-  :bootstrap_options => {
-    :template => VM_TEMPLATE
-  }
-)
-
-one_auth = Chef::Provisioning::OpenNebulaDriver.get_onelib(
-  :driver_url => DRIVER_URL
-).client.one_auth
+username = one_auth.split(':').first
+machine 'RSpec-test-vm' do
+  machine_options MACHINE_OPTIONS.merge(
+    :bootstrap_options => {
+      :template_name => 'RSpec-test-template',
+      :template_options => {
+        'NIC' => {
+          'NETWORK' => 'RSpec-test-vnet',
+          'NETWORK_UNAME' => username
+        }
+      },
+      :mode => '600'
+    }
+  )
+  action :allocate
+end