+++ spec/recipes/OneFlow/service_action_hold.rb
@@ -12,20 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require 'chef/provisioning/opennebula_driver'
-require 'fileutils'
-
-FileUtils.mkdir_p(CHEF_REPO_PATH.chomp('/'))
-with_chef_local_server :chef_repo_path => CHEF_REPO_PATH.chomp('/')
-with_driver DRIVER_URL
-with_flow_url ONE_FLOW_URL.chomp('/') unless ONE_FLOW_URL.nil?
-
-with_machine_options MACHINE_OPTIONS.merge(
-  :bootstrap_options => {
-    :template => VM_TEMPLATE
-  }
-)
-
-one_auth = Chef::Provisioning::OpenNebulaDriver.get_onelib(
-  :driver_url => DRIVER_URL
-).client.one_auth
+one_flow_service 'RSpec-role-action-service' do
+  role 'RSpec_hold'
+  # The following setting will ignore the fact the feature is not fully tested and run it anyway
+  #override_failsafe true
+  action :hold
+end